Change array NaCl.Sodium len values to ints.
Java arrays are always indexed by ints and thus limited to Integer.MAX
elements, in theory. I was worried that the array length values were
defined as longs for a performance reason with jnr-ffi but in testing I
couldn't see any difference.

It is better to define them as ints here to be consistent with the Java
language, I think.